+++ py/desimodel/weather.py
@@ -27,10 +27,16 @@ survey strategy studies.
 """
 from __future__ import print_function, division
 
+import os
+import datetime
+import calendar
+
 import numpy as np
 
 import scipy.interpolate
 import scipy.special
+
+import astropy.table
 
 
 def whiten_transforms_from_cdf(x, cdf):
@@ -356,3 +362,62 @@ def sample_transp(n_sample, dt_sec=180., gen=None):
     transp_grid, pdf_grid = get_transp_pdf()
     return sample_timeseries(
         transp_grid, pdf_grid, _transp_psd, n_sample, dt_sec, gen)
+
+
+def dome_closed_fractions(start_date, stop_date, replay=np.arange(2007,2018)):
+    """Return dome-closed fractions for each night of the survey.
+
+    Years can be replayed in any order.  If the number of years to replay is less
+    than the survey duration, they are repeated.
+
+    Parameters
+    ----------
+    start_date : datetime.date or None
+        Survey starts on the evening of this date. Use the ``first_day``
+        config parameter if None (the default).
+    stop_date : datetime.date or None
+        Survey stops on the morning of this date. Use the ``last_day``
+        config parameter if None (the default).
+    replay : list
+        List of integer years to replay in the range 2007-2017.
+
+    Returns
+    -------
+    numpy array
+        1D array of N probabilities between 0-1, where N is the number of nights
+        spanned by the start and stop dates.
+    """
+    # Check the inputs.
+    num_nights = (stop_date - start_date).days
+    if num_nights <= 0:
+        raise ValueError('Expected start_date < stop_date.')
+    replay = np.asarray(replay)
+    if not np.all((replay >= 2007) & (replay <= 2017)):
+        raise ValueError('Invalid replay years.')
+    # Load tabulated daily weather history.
+    DESIMODEL = os.environ['DESIMODEL']
+    path = os.path.join(DESIMODEL, 'data', 'weather', 'daily-2007-2017.dat')
+    t = astropy.table.Table.read(path, format='ascii')
+    if not len(t) % 365 == 0:
+        raise ValueError('Invalid weather history length.')
+    lostfrac = t['lostfrac'].data
+    if not np.all((lostfrac >= 0) & (lostfrac <= 1)):
+        raise ValueError('Invalid weather history data.')
+    # Replay the specified years (with wrap-around if necessary),
+    # overlaid on the actual survey dates.
+    probs = np.zeros(num_nights)
+    start = start_date
+    for year_num, year in enumerate(range(start_date.year, stop_date.year + 1)):
+        first = datetime.date(year=year, month=1, day=1)
+        stop = datetime.date(year=year + 1, month=1, day=1)
+        if stop > stop_date:
+            stop = stop_date
+        n = (stop - start).days
+        if calendar.isleap(year):
+            n -= 1
+        idx = (start - start_date).days
+        offset = 365 * (replay[year_num % len(replay)] - 2007)
+        jdx = offset + (start - first).days
+        probs[idx:idx + n] = lostfrac[jdx:jdx + n]
+        start = stop
+    return probs